+++ PHP/CodeCoverage/Filter.php
@@ -76,41 +76,14 @@ class PHP_CodeCoverage_Filter
      */
     public function __construct()
     {
-        $functions = array(
-          'file_iterator_autoload',
-          'php_codecoverage_autoload',
-          'php_invoker_autoload',
-          'php_timer_autoload',
-          'php_tokenstream_autoload',
-          'phpunit_autoload',
-          'phpunit_dbunit_autoload',
-          'phpunit_mockobject_autoload',
-          'phpunit_selenium_autoload',
-          'phpunit_story_autoload',
-          'text_template_autoload'
-        );
-
-        foreach ($functions as $function) {
-            if (function_exists($function)) {
-                $this->addFilesToBlacklist($function());
-            }
-        }
-
-        $file = stream_resolve_include_path(
-          'SymfonyComponents/YAML/sfYaml.php'
-        );
-
-        if ($file) {
-            $this->addFileToBlacklist($file);
-        }
-
-        $file = stream_resolve_include_path(
-          'SymfonyComponents/YAML/sfYamlDumper.php'
-        );
-
-        if ($file) {
-            $this->addFileToBlacklist($file);
-        }
+        $this->addDirectoryContainingClassToBlacklist('File_Iterator');
+        $this->addDirectoryContainingClassToBlacklist('PHP_CodeCoverage');
+        $this->addDirectoryContainingClassToBlacklist('PHP_Invoker');
+        $this->addDirectoryContainingClassToBlacklist('PHP_Timer');
+        $this->addDirectoryContainingClassToBlacklist('PHP_Token');
+        $this->addDirectoryContainingClassToBlacklist('PHPUnit_Framework_TestCase', 2);
+        $this->addDirectoryContainingClassToBlacklist('Text_Template');
+        $this->addDirectoryContainingClassToBlacklist('Symfony\Component\Yaml\Yaml');
     }
 
     /**
@@ -337,5 +310,26 @@ class PHP_CodeCoverage_Filter
     public function hasWhitelist()
     {
         return !empty($this->whitelistedFiles);
+    }
+
+    /**
+     * @param string  $className
+     * @param integer $parent
+     * @since Method available since Release 1.2.3
+     */
+    protected function addDirectoryContainingClassToBlacklist($className, $parent = 1)
+    {
+        if (!class_exists($className)) {
+            return;
+        }
+
+        $reflector = new ReflectionClass($className);
+        $directory = $reflector->getFileName();
+
+        for ($i = 0; $i < $parent; $i++) {
+            $directory = dirname($directory);
+        }
+
+        $this->addDirectoryToBlacklist($directory);
     }
 }